requirements.txt
couchdb
genologics @ git+https://github.com/SciLifeLab/genologics.git
genologics_sql @ git+https://github.com/SciLifeLab/genologics_sql.git
markdown
pyyaml
requests
setuptools
six
sqlalchemy
statusdb @ git+https://github.com/NationalGenomicsInfrastructure/statusdb.git